Backend/requirements.txt
google-cloud-language
fastapi[standard]
google-cloud-storage
google.cloud.videointelligence
google.cloud.vision
dotenv
mysql-connector-python